+++ packages/messenger-widget/src/components/DM3/DM3.tsx
@@ -1,37 +1,14 @@
-/* eslint-disable no-console */
-import { useContext, useEffect, useState } from 'react';
+import { useContext, useEffect } from 'react';
 import { AuthContext } from '../../context/AuthContext';
 import { Dm3Props } from '../../interfaces/config';
-import { GlobalContext } from '../../utils/context-utils';
-import { ConnectionType } from '../../utils/enum-type-utils';
 import Dashboard from '../../views/Dashboard/Dashboard';
 import { SignIn } from '../SignIn/SignIn';
 import { DM3ConfigurationContext } from '../../context/DM3ConfigurationContext';
 
 function DM3(props: Dm3Props) {
-    // fetches context storage
-    const { dispatch } = useContext(GlobalContext);
+    const { setDm3Configuration } = useContext(DM3ConfigurationContext);
 
-    const { setDm3Configuration, setScreenWidth, screenWidth } = useContext(DM3ConfigurationContext);
-
-    const { isLoggedIn, account, deliveryServiceToken } =
-        useContext(AuthContext);
-
-    // handles changes of state on connection to the account
-    //TODO refactor to useAuth
-    useEffect(() => {
-        if (props.config.connectionStateChange) {
-            props.config.connectionStateChange(isLoggedIn);
-        }
-    }, [isLoggedIn]);
-
-    // TODO what is this doing?
-    useEffect(() => {
-        dispatch({
-            type: ConnectionType.SetDefaultServiceUrl,
-            payload: props.config.defaultServiceUrl,
-        });
-    }, [props.config.defaultServiceUrl]);
+    const { isLoggedIn } = useContext(AuthContext);
 
     // updates rainbow kit provider height to 100% when rendered
     useEffect(() => {
@@ -42,17 +19,6 @@ function DM3(props: Dm3Props) {
 
         // sets the DM3 confguration provided from props
         setDm3Configuration(props.dm3Configuration);
-    }, []);
-
-    // This handles the responsive check of widget
-    useEffect(() => {
-        const handleResize = () => {
-            setScreenWidth(window.innerWidth);
-        };
-        window.addEventListener('resize', handleResize);
-        return () => {
-            window.removeEventListener('resize', handleResize);
-        };
     }, []);
 
     return (
