+++ packages/messenger-widget/src/context/DM3ConfigurationContext.tsx
@@ -5,13 +5,11 @@ import { useDm3Configuration } from '../hooks/configuration/useDM3Configuration'
 export type DM3ConfigurationContextType = {
     setDm3Configuration: (configuration: DM3Configuration) => void;
     dm3Configuration: DM3Configuration;
-    screenWidth: number;
-    setScreenWidth: (width: number) => void;
 };
 
 export const DM3ConfigurationContext =
     React.createContext<DM3ConfigurationContextType>({
-        setDm3Configuration: (configuration: DM3Configuration) => { },
+        setDm3Configuration: (configuration: DM3Configuration) => {},
         dm3Configuration: {
             defaultContact: '',
             defaultServiceUrl: '',
@@ -28,8 +26,6 @@ export const DM3ConfigurationContext =
             showAlways: true,
             showContacts: true,
         },
-        screenWidth: window.innerWidth,
-        setScreenWidth: (width: number) => { }
     });
 
 export const DM3ConfigurationContextProvider = ({
@@ -37,16 +33,13 @@ export const DM3ConfigurationContextProvider = ({
 }: {
     children?: any;
 }) => {
-    const { dm3Configuration, setDm3Configuration, screenWidth,
-        setScreenWidth } = useDm3Configuration();
+    const { dm3Configuration, setDm3Configuration } = useDm3Configuration();
 
     return (
         <DM3ConfigurationContext.Provider
             value={{
                 dm3Configuration,
                 setDm3Configuration,
-                screenWidth,
-                setScreenWidth
             }}
         >
             {children}
