--- packages/messenger-widget/src/interfaces/web3.ts
@@ -1,23 +1,0 @@
-import { ethers } from 'ethers';
-import { Socket } from 'socket.io-client';
-import { DefaultEventsMap } from 'socket.io/dist/typed-events';
-import { Account } from '@dm3-org/dm3-lib-profile';
-import { ConnectionState } from '../utils/enum-type-utils';
-
-declare global {
-    interface Window {
-        ethereum?: any;
-    }
-}
-
-export interface Connection {
-    socket?: Socket<DefaultEventsMap, DefaultEventsMap>;
-    defaultServiceUrl: string;
-}
-
-export interface SignInProps {
-    hideStorageSelection: boolean;
-    miniSignIn: boolean;
-    defaultStorageLocation: any | undefined;
-    signInImage: string;
-}
